requirements.txt
# Install transformers from source
git+https://github.com/huggingface/transformers

# Core dependencies
torch>=2.0.0
numpy==1.26.0
timm
torchvision
Pillow

# Training tools
datasets
accelerate
albumentations==1.4.6
torchmetrics

# Optional utilities
tensorboard
tqdm
matplotlib
